+++ src/main/java/com/idle/fmd/domain/board/entity/BoardEntity.java
@@ -1,7 +1,6 @@
 package com.idle.fmd.domain.board.entity;
 
 import com.idle.fmd.domain.board.dto.BoardCreateDto;
-import com.idle.fmd.domain.file.entity.FileEntity;
 import com.idle.fmd.domain.user.entity.UserEntity;
 import com.idle.fmd.global.common.entity.BaseTimeEntity;
 import jakarta.persistence.*;
