+++ src/main/java/com/idle/fmd/domain/board/repo/FileRepository.java
@@ -1,6 +1,6 @@
-package com.idle.fmd.domain.file.repo;
+package com.idle.fmd.domain.board.repo;
 
-import com.idle.fmd.domain.file.entity.FileEntity;
+import com.idle.fmd.domain.board.entity.FileEntity;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface FileRepository extends JpaRepository<FileEntity, Long> {
