+++ src/main/java/com/idle/fmd/domain/board/service/BoardService.java
@@ -5,13 +5,13 @@ import com.idle.fmd.domain.board.dto.BoardResponseDto;
 import com.idle.fmd.domain.board.dto.BoardUpdateDto;
 import com.idle.fmd.domain.board.entity.BoardEntity;
 import com.idle.fmd.domain.board.repo.BoardRepository;
-import com.idle.fmd.domain.file.entity.FileEntity;
-import com.idle.fmd.domain.file.repo.FileRepository;
+import com.idle.fmd.domain.board.entity.FileEntity;
+import com.idle.fmd.domain.board.repo.FileRepository;
 import com.idle.fmd.domain.comment.entity.CommentEntity;
 import com.idle.fmd.domain.comment.repo.CommentRepository;
 import com.idle.fmd.domain.user.entity.UserEntity;
 import com.idle.fmd.domain.user.repo.UserRepository;
-import com.idle.fmd.global.common.utils.FileHandler;
+import com.idle.fmd.global.utils.FileHandler;
 import com.idle.fmd.global.error.exception.BusinessException;
 import com.idle.fmd.global.error.exception.BusinessExceptionCode;
 import lombok.RequiredArgsConstructor;
