+++ src/main/java/com/idle/fmd/global/auth/jwt/JwtTokenFilter.java
@@ -3,7 +3,7 @@ package com.idle.fmd.global.auth.jwt;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import com.idle.fmd.domain.user.dto.UserLoginResponseDto;
 import com.idle.fmd.domain.user.entity.CustomUserDetails;
-import com.idle.fmd.global.common.utils.RedisUtil;
+import com.idle.fmd.global.utils.RedisUtil;
 import jakarta.servlet.FilterChain;
 import jakarta.servlet.ServletException;
 import jakarta.servlet.http.HttpServletRequest;
